Created the class files for "Family", "User", ""AppSettings", and event. Also refactored the code so all class files are in the same folder.
