+++ app/src/main/java/com/example/famplanapp/NewTaskComponents.kt
@@ -41,6 +41,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.unit.toSize
+import com.example.famplanapp.classes.Task
 import java.time.LocalDateTime
 import java.util.Calendar
