+++ app/src/main/java/com/example/famplanapp/Tasks.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
+import com.example.famplanapp.classes.Task
 import java.time.format.DateTimeFormatter
-import java.util.Date
 
 var tasksList = mutableListOf<Task>()
