+++ app/src/main/java/com/example/famplanapp/classes/AppSettings.kt
@@ -1,0 +1,4 @@
+package com.example.famplanapp.classes
+
+class AppSettings(var sharedBudget: Boolean = false, var notificationSettings: String = "Push", var usePreferredName: Boolean = false) {
+}
