+++ app/src/main/java/com/example/famplanapp/classes/Event.kt
@@ -1,0 +1,7 @@
+package com.example.famplanapp.classes
+
+import java.util.Date
+
+class Event(var id: Int = 0, var title: String = "", var date: Date = Date(), var invitedUsers: List<User>) {
+
+}
