+++ app/src/main/java/com/example/famplanapp/classes/Family.kt
@@ -1,0 +1,4 @@
+package com.example.famplanapp.classes
+
+class Family(var id: Int, var settings: AppSettings, var users: List<User>) {
+}
