+++ app/src/main/java/com/example/famplanapp/classes/Poll.kt
@@ -1,0 +1,12 @@
+package com.example.famplanapp.classes
+
+import com.example.famplanapp.PollOption
+import java.time.LocalDateTime
+
+data class Poll(
+    val id: Int,
+    val owner: User, // should be User object eventually
+    val subject: String,
+    val options: List<PollOption>,
+    val deadline: LocalDateTime
+)
